+++ terraform/aws_load_balancer_controller.tf
@@ -135,4 +135,58 @@ resource "helm_release" "aws_load_balancer_controller" {
     name  = "serviceAccount.create"
     value = "false"
   }
+
+  # set {
+  #   name  = "alb\\.ingress\\.kubernetes\\.io/certificate-arn"
+  #   value = aws_acm_certificate.ice01.arn
+  # }
+}
+
+resource "kubernetes_ingress_v1" "ice01" {
+  metadata {
+    name = "hello-kubernetes"
+    labels = {
+      "app.kubernetes.io/name"       = "hello-kubernetes"
+      "app.kubernetes.io/managed-by" = "terraform"
+    }
+    annotations = {
+      "kubernetes.io/ingress.class"      = "alb"
+      "alb.ingress.kubernetes.io/scheme" = "internet-facing"
+      # "alb.ingress.kubernetes.io/listen-ports" = "[{'HTTP': 80}, {'HTTPS': 443}]"
+      "alb.ingress.kubernetes.io/certificate-arn" = aws_acm_certificate.ice01.arn
+    }
+  }
+  spec {
+    default_backend {
+      service {
+        name = "hello-kubernetes"
+        port { number = 8080 }
+      }
+    }
+    rule {
+      http {
+        path {
+          path      = "/"
+          path_type = "Prefix"
+          backend {
+            service {
+              name = "hello-kubernetes"
+              port { number = 8080 }
+            }
+          }
+        }
+      }
+    }
+  }
+
+}
+
+resource "aws_route53_record" "ice01_naked" {
+  name    = "origin.icekernelcloud01.com"
+  type    = "CNAME"
+  zone_id = data.aws_route53_zone.zone.zone_id
+  ttl     = 360
+  records = [
+    kubernetes_ingress_v1.ice01.status[0].load_balancer[0].ingress[0].hostname,
+  ]
 }