+++ terraform/variables.tf
@@ -3,13 +3,17 @@ variable "cluster_name" {
   default = "ice01"
 }
 
-data "aws_region" "current" {}
-
-data "aws_eks_cluster" "cluster" {
-  name = module.eks.cluster_id
+variable "domain_name" {
+  type    = string
+  default = "icekernelcloud01.com"
 }
 
-data "aws_eks_cluster_auth" "cluster" {
-  name = module.eks.cluster_id
+variable "subdomain" {
+  default = "www"
 }
 
+variable "tags" {
+  default = {
+    ManagedBy = "terraform"
+  }
+}