+++ CHANGELOG.md
@@ -1,4 +1,13 @@
 
+# Thu Apr 28 22:05:44 UTC 2022
+
+added:
+
+- route53 provisitioned hostnames
+- acm signed certificate
+- ingress setup through terraform
+- cloudfront
+
 # Tue Apr 26 21:47:17 UTC 2022
 
 reorganized most of the code structure to add more things from the todo list.